Merge gbtami/pychess-variants into liantichess

+++ server/generate_highscore.py
@@ -3,23 +3,36 @@ from decimal import Decimal
 from const import VARIANTS, HIGHSCORE_MIN_GAMES
 
 
-async def generate_highscore(db):
+async def generate_highscore(db, one_variant=None):
+    limit = 50
     hs = []
-    for variant in VARIANTS:
+    variants = VARIANTS if one_variant is None else (one_variant,)
+
+    for variant in variants:
         # print(variant)
         d = "perfs.%s.gl.d" % variant
         r = "perfs.%s.gl.r" % variant
         nb = "perfs.%s.nb" % variant
-        filt = {d: {"$lt": 350}, "enabled": {"$ne": False}, nb: {"$gte": HIGHSCORE_MIN_GAMES}}
+        filt = {
+            d: {"$lt": 350},
+            "enabled": {"$ne": False},
+            nb: {"$gte": HIGHSCORE_MIN_GAMES},
+        }
 
         scores = {}
-        cursor = db.user.find(filt, sort=[(r, -1)], limit=10)
+        cursor = db.user.find(filt, sort=[(r, -1)], limit=limit)
         async for doc in cursor:
             scores[doc["_id"]] = int(round(Decimal(doc["perfs"][variant]["gl"]["r"]), 0))
 
-        hs.append({"_id": variant, "scores": scores})
+        if one_variant is None:
+            hs.append({"_id": variant, "scores": scores})
+        else:
+            hs = scores
 
-    await db.highscore.drop()
-    # bulk insert to highscore
-    if len(hs) > 0:
-        await db.highscore.insert_many(hs)
+    if one_variant is None:
+        await db.highscore.drop()
+        # bulk insert to highscore
+        if len(hs) > 0:
+            await db.highscore.insert_many(hs)
+
+    return hs
